Production bundling
This creates a gulp target for `production`, allowing creation of a
jspm bundle containing all javascript and css. This also triggers
compilation of an index.html for the public directory that depends
solely on the new bundle.

Both the dev environment and the production bundle work using the
same setup.