Refactors AnalyseProject -> AnalysePull/PushAnalysisComments (#67)
Splits the AnalyseProject job into two separate jobs, one
responsible for creating the PullAnalysis record and the other for
sending the comments to Github.

PullAnalysis records and now created for multiple pushed to the
same pull, meaning that analysis comments can be duplicated on
later analyses. PushAnalysisComments ensures that only new comments
are sent to Github.