Fix ssh auth take 2! (#84)
The previous faux-fix removed the publickey from the fields used
to initialize an Rugged::Credentials::SshKey, which worked on my
local machine as my libgcrypt can determine publickeys from
privatekeys.

Ubuntu on prod libgcrypt does not have this functionality, which
raised exceptions. Having now tried to find the workaround, it
looks like libssh2 hangs on public keys that don't have the comment.

`Project.generate_keypair!` never generated comments, and as a
result the keys would fail when passed to Rugged.