Apriori w/ TID
This is a serviceable algorithm for identifying frequent pattern
sets in version control history, and for the first 5,000 changesets
of rails be run with min_support=10 in 8s.

Caveat is that we have to default to max_items=10 to avoid
combinatorial processing increases from subset generation of larger
changesets.

Next candidate will be fp growth, as this is reportedly far faster
than Apriori.